+++ tasks/helpers.js
@@ -1,9 +1,8 @@
-var gutil = require('gulp-util');
 var exec = require('child_process').exec;
 
 module.exports = {
   childProcess: function(command, arguments, cb) {
-    gutil.log(gutil.colors.cyan('running command: '), command + ' ' + arguments.join(' '));
+    console.log('running command: ' + command + ' ' + arguments.join(' '));
     exec(command + ' ' + arguments.join(' '), cb);
   }
 };