+++ web/app/application.jsx
@@ -1,2 +1,7 @@
-//load some modules that will load the startup code for our application
+//static resources
+require('./../index.html');
+require('./misc/ua-parser.js');
+
+require('./styles/main.scss');
+
 require('./router.jsx');