+++ web/app/pages/desktop/desktop.component.jsx
@@ -1,41 +1,36 @@
-var React = require('react/addons');
-var menuStore = require('../../stores/menu.store');
-var userStore = require('../../stores/user.store');
+var React = require('react');
+var reactRedux = require('react-redux');
+var actions = require('../../store/actions');
+var usersRepository = require('../../repositories/users.repository');
 
 var desktop = {};
 
 desktop.displayName = 'Desktop';
 
-desktop.getInitialState = function desktopComponentGetInitialState() {
-  return {
-    user: null
-  };
+desktop.contextTypes = {
+  store: React.PropTypes.object
 };
 
 desktop.componentDidMount = function desktopComponentComponentDidMount() {
-  menuStore.update({
-    menuName: 'desktop'
-  });
+  this.context.store.dispatch(actions.menu.setActive('desktop'));
 };
 
 desktop.onClickGetUser = function desktopComponentOnClickGetUser() {
-  userStore.getUser(123).then(function desktopComponentOnGetUserSuccess(user) {
-    this.setState({
-      user: user
-    });
-  }.bind(this));
+  setTimeout(function() {
+    usersRepository.getUser(123);
+  }, 2000);
 };
 
 desktop.renderUserData = function desktopComponentRenderUserData() {
   var userData = null;
 
-  if (this.state.user) {
+  if (this.props.user) {
     userData = (
       <ul>
-        <li className="user-id">{this.state.user.id}</li>
-        <li className="user-username">{this.state.user.username}</li>
-        <li className="user-first-name">{this.state.user.firstName}</li>
-        <li className="user-last-name">{this.state.user.lastName}</li>
+        <li className="user-id">{this.props.user.id}</li>
+        <li className="user-username">{this.props.user.username}</li>
+        <li className="user-first-name">{this.props.user.firstName}</li>
+        <li className="user-last-name">{this.props.user.lastName}</li>
       </ul>
     );
   }
@@ -48,6 +43,7 @@ desktop.render = function desktopComponentRender() {
     <div className="p-desktop">
       <h1 id="test" className="test">{window.i18n['desktop/desktop'].header()}</h1>
       <div>
+        <img src={require('../../images/user.png')} />
         <button className="load-user-data" onClick={this.onClickGetUser}>Get User Data</button>
         {this.renderUserData()}
       </div>
@@ -55,4 +51,10 @@ desktop.render = function desktopComponentRender() {
   );
 };
 
-module.exports = React.createClass(desktop);
+var mapStateToProps = function(state) {
+  return {
+    user: state.users.getIn(['activeUser']) ? state.users.getIn(['activeUser']).toJS() : null
+  };
+};
+
+module.exports = reactRedux.connect(mapStateToProps)(React.createClass(desktop));