+++ web/app/pages/desktop/desktop.hooks.js
@@ -1,0 +1,9 @@
+var store = require('../../store/store');
+var actions = require('../../store/actions');
+
+module.exports = {
+  onEnter: function(nextState, replace, callback) {
+    store.dispatch(actions.users.clearActive());
+    callback();
+  }
+};