+++ web/app/pages/desktop/module.jsx
@@ -1,4 +1,4 @@
-var React = require('react/addons');
+var React = require('react');
 var Router = require('react-router');
 var Route = Router.Route;
 
@@ -7,11 +7,36 @@ var PreventDoubleClick = require('./prevent-double-click.component.jsx');
 var WithResolves = require('./with-resolves.component.jsx');
 var WithParam = require('./with-param.component.jsx');
 
+var withResolvesHooks = require('./with-resolves.hooks');
+var desktopHooks = require('./desktop.hooks');
+
 module.exports = {
   routes: [
-    <Route key="1" name="desktop" path="/desktop" handler={Desktop} />,
-    <Route key="2" name="prevent-double-click" path="/prevent-double-click" handler={PreventDoubleClick} />,
-    <Route key="3" name="with-resolves" path="/with-resolves" handler={WithResolves} />,
-    <Route key="4" name="with-param" path="/with-param/:p1?" handler={WithParam} />
+    <Route
+      key="1"
+      name="desktop"
+      path="/desktop"
+      component={Desktop}
+      onEnter={desktopHooks.onEnter}
+    />,
+    <Route
+      key="2"
+      name="prevent-double-click"
+      path="/prevent-double-click"
+      component={PreventDoubleClick}
+    />,
+    <Route
+      key="3"
+      name="with-resolves"
+      path="/with-resolves"
+      component={WithResolves}
+      onEnter={withResolvesHooks.onEnter}
+    />,
+    <Route
+      key="4"
+      name="with-param"
+      path="/with-param(/:p1)"
+      component={WithParam}
+    />
   ]
 };