+++ web/app/react/components/application.component.jsx
@@ -1,6 +1,6 @@
-var React = require('react/addons');
+var React = require('react');
+
 var Header = require('./header.component.jsx');
-var RouteHandler = require('react-router').RouteHandler;
 
 //TODO: remove: for testing assets rewrite process
 var removeMe = '/app/svg/remove-me.svg#icon-small';
@@ -13,7 +13,7 @@ application.render = function applicationRender() {
   return (
     <div className="application">
       <Header />
-      <RouteHandler routerState={this.props.routerState} />
+      {this.props.children}
     </div>
   );
 };