+++ web/app/router.jsx
@@ -1,6 +1,5 @@
-var React = require('react/addons');
+var React = require('react');
+var ReactDOM = require('react-dom');
 var Router = require('react-router');
 
-Router.run(require('./routes.jsx'), Router.HistoryLocation, function routerRun(Handler, routerState) {
-  React.render(<Handler routerState={routerState} />, document.querySelector('.react-bootstrap-element'));
-});
+ReactDOM.render(require('./routes.jsx'), document.querySelector('.react-bootstrap-element'));