+++ web/app/routes.jsx
@@ -1,17 +1,28 @@
-var React = require('react/addons');
+var React = require('react');
+var reactRouter = require('react-router');
 var applicationReact = require('./react/index');
+var reactRedux = require('react-redux');
+var reactRouterRedux = require('react-router-redux');
+var store = require('./store/store');
+
+var history = reactRouterRedux.syncHistoryWithStore(reactRouter.browserHistory, store);
+
 var Application = applicationReact.components.Application;
 var Desktop = require('./pages/desktop/desktop.component.jsx');
-var Router = require('react-router');
-var Route = Router.Route;
-var DefaultRoute = Router.DefaultRoute;
-var NotFoundRoute = Router.NotFoundRoute;
+var IndexRoute = reactRouter.IndexRoute;
 var NotFound = applicationReact.components.NotFound;
+var Provider = reactRedux.Provider;
+var Route = reactRouter.Route;
+var Router = reactRouter.Router;
 
 module.exports = (
-  <Route handler={Application}>
-    <DefaultRoute handler={Desktop} />
-    {require('./pages/desktop/module.jsx').routes}
-    <NotFoundRoute handler={NotFound} />
-  </Route>
+  <Provider store={store}>
+    <Router history={history}>
+      <Route path="/" component={Application}>
+        <IndexRoute component={Desktop} />
+        {require('./pages/desktop/module.jsx').routes}
+        <Route path="*" component={NotFound} />
+      </Route>
+    </Router>
+  </Provider>
 );