+++ web/app/store/actions.js
@@ -1,0 +1,5 @@
+module.exports = {
+  menu: require('./actions/menu.actions'),
+  preventDoubleClick: require('./actions/prevent-double-click.actions'),
+  users: require('./actions/users.actions')
+};