+++ web/app/store/actions/menu.actions.js
@@ -1,0 +1,8 @@
+module.exports = {
+  setActive: function(menuName) {
+    return {
+      type: 'Menu::setActive',
+      menuName: menuName
+    };
+  }
+};