+++ web/app/store/actions/prevent-double-click.actions.js
@@ -1,0 +1,13 @@
+module.exports = {
+  enable: function() {
+    return {
+      type: 'PreventDoubleClick::enable'
+    };
+  },
+
+  disable: function() {
+    return {
+      type: 'PreventDoubleClick::disable'
+    };
+  }
+};