+++ web/app/store/actions/users.actions.js
@@ -1,0 +1,14 @@
+module.exports = {
+  setActive: function(user) {
+    return {
+      type: 'Users::setActive',
+      user: user
+    };
+  },
+
+  clearActive: function() {
+    return {
+      type: 'Users::clear'
+    };
+  }
+};