+++ web/app/stores/store-generator.js
@@ -1,8 +1,0 @@
-var EventEmitter = require('eventemitter3');
-var objectAssign = require('object-assign');
-
-module.exports = function storeGenerator(store) {
-  objectAssign(store, EventEmitter.prototype);
-
-  return store;
-};